Boston House Price Data (Original)

Original data, which come from a study by Harrison Jr and Rubinfeld (1978), examining the association between median house prices in a particular community with various community characteristics. See bostonhouseprice2 for the corrected version, with additional variables.

Usage

bostonhouseprice

Format

A data frame with 506 rows and 9 variables:

crime

crimes committed per capita

rooms

average number of rooms per house

radial

index of accessibility to radial highways

stratio

average student-teacher ratio of schools in the community

lowstat

percentage of the population that are "lower status"

lnox

log(annual average nitrogen oxide concentration (pphm))

lproptax

log(property tax per $1000)

ldist

log(weighted distances to five employment centres in the Boston region)

lprice

log(median house price ($))

Source

https://CRAN.R-project.org/package=wooldridge

References

Harrison Jr, D. and Rubinfeld, D. L. (1978). Hedonic housing prices and the demand for clean air. Journal of environmental economics and management, 5(1):81-102.

Wooldridge, J. M. (2015). Introductory econometrics: A modern approach. Cengage learning.