Boston House Price Data (Corrected Version)

Corrected data, which come from a study by Harrison Jr and Rubinfeld (1978), examining the association between median house prices in a particular community with various community characteristics. See bostonhouseprice for the original version.

Usage

bostonhouseprice2

Format

A data frame with 506 rows and 13 variables:

crim

per capita crime rate by town

zn

proportion of residential land zoned for lots over 25,000 sq.ft

indus

proportion of non-retail business acres per town

rm

average number of rooms per dwelling

age

proportion of owner-occupied units built prior to 1940

rad

index of accessibility to radial highways

ptratio

pupil-teacher ratio by town

lnox

log(nitric oxides concentration (parts per 10 million))

ldis

log(weighted distances to five Boston employment centres)

ltax

log(full-value property-tax rate per USD 10,000)

llstat

log(percentage of lower status of the population)

chast

Charles River dummy variable (=1 if tract bounds river; 0 otherwise)

lcmedv

log(corrected median value of owner-occupied homes in USD 1000's)

Source

https://CRAN.R-project.org/package=mlbench

References

Harrison Jr, D. and Rubinfeld, D. L. (1978). Hedonic housing prices and the demand for clean air. Journal of environmental economics and management, 5(1):81-102.

Leisch F, Dimitriadou E (2021). mlbench: Machine Learning Benchmark Problems. R package version 2.1-3.