Diabetes Data

Data relating to a study of disease progression one year after baseline.

Usage

diabetes

Format

A data frame with 442 rows and 11 variables:

AGE

age of the patient

SEX

sex of the patient

BMI

body mass index of the patient

BP

blood pressure of the patient

S1

blood serum measurement 1

S2

blood serum measurement 2

S3

blood serum measurement 3

S4

blood serum measurement 4

S5

blood serum measurement 5

S6

blood serum measurement 6

Y

quantitative measure of disease progression one year after baseline

Source

https://CRAN.R-project.org/package=lars

References

Efron, B., Hastie, T., Johnstone, I., Tibshirani, R., et al. (2004). Least angle regression. The Annals of Statistics.