Predict smoothic

predict method class “smoothic”

Usage

# S3 method for class 'smoothic'
predict(object, newdata, ...)

Arguments

object

an object of class “smoothic” which is the result of a call to smoothic.

newdata

new data object

...

further arguments passed to or from other methods.

Value

a matrix containing the predicted values for the location mu and scale s

Author

Meadhbh O'Neill

Examples

# Sniffer Data --------------------
# MPR Model ----
results <- smoothic(
  formula = y ~ .,
  data = sniffer,
  family = "normal",
  model = "mpr"
)
predict(results)
#>           mu        s
#> 1   23.85742 1.844882
#> 2   26.20599 2.820692
#> 3   27.57423 3.249511
#> 4   26.97700 3.158825
#> 5   26.36898 2.518755
#> 6   23.07260 2.008381
#> 7   22.01057 1.897847
#> 8   22.37036 1.952332
#> 9   25.87346 3.070670
#> 10  24.81508 2.984975
#> 11  25.21216 3.070670
#> 12  24.99629 3.070670
#> 13  27.58472 3.342801
#> 14  27.60730 3.342801
#> 15  27.48072 3.342801
#> 16  28.30795 3.639049
#> 17  21.45270 2.008381
#> 18  21.29305 1.952332
#> 19  20.59503 1.844882
#> 20  20.96633 1.952332
#> 21  32.01594 3.537492
#> 22  32.50231 3.639049
#> 23  33.28176 3.850994
#> 24  33.50812 3.961552
#> 25  33.50812 3.961552
#> 26  33.50812 3.961552
#> 27  31.98537 3.961552
#> 28  31.69593 3.961552
#> 29  31.46957 3.850994
#> 30  32.52009 3.961552
#> 31  32.43080 3.961552
#> 32  32.34151 3.961552
#> 33  31.98172 3.850994
#> 34  32.20808 3.961552
#> 35  31.62136 3.961552
#> 36  31.87772 3.850994
#> 37  31.52842 3.850994
#> 38  30.59661 3.743522
#> 39  31.40971 3.850994
#> 40  31.63607 3.961552
#> 41  32.17444 4.075283
#> 42  32.40080 4.192280
#> 43  31.60243 4.075283
#> 44  31.25735 3.961552
#> 45  24.01929 2.008381
#> 46  23.38056 2.008381
#> 47  25.59711 2.186369
#> 48  21.95337 2.125353
#> 49  20.59141 1.952332
#> 50  30.08767 3.639049
#> 51  30.73733 3.961552
#> 52  30.75205 3.961552
#> 53  31.30513 4.075283
#> 54  31.83986 4.075283
#> 55  31.39442 4.075283
#> 56  30.97841 4.075283
#> 57  42.03115 4.436446
#> 58  41.22855 4.436446
#> 59  42.82031 3.537492
#> 60  43.72577 3.961552
#> 61  50.99575 9.002164
#> 62  52.09085 9.799960
#> 63  52.09085 9.799960
#> 64  52.09085 9.799960
#> 65  50.65010 9.260605
#> 66  51.19212 9.799960
#> 67  32.54038 4.968269
#> 68  29.98616 4.192280
#> 69  30.54347 4.192280
#> 70  30.07545 4.192280
#> 71  23.83943 2.518755
#> 72  22.91869 2.380131
#> 73  41.00469 4.968269
#> 74  39.96523 4.694833
#> 75  37.75597 4.563811
#> 76  39.01757 5.110902
#> 77  45.91490 8.506717
#> 78  47.08093 9.002164
#> 79  47.39659 9.260605
#> 80  31.16748 4.192280
#> 81  31.50892 4.192280
#> 82  31.03041 4.075283
#> 83  31.07819 4.192280
#> 84  30.51040 4.075283
#> 85  30.18367 4.075283
#> 86  30.56240 4.075283
#> 87  30.56240 4.075283
#> 88  30.65169 4.075283
#> 89  30.78876 4.192280
#> 90  30.11274 4.192280
#> 91  30.92583 4.312635
#> 92  31.34606 4.192280
#> 93  32.09244 4.312635
#> 94  31.16748 4.192280
#> 95  30.69947 4.192280
#> 96  30.41004 4.192280
#> 97  30.20203 4.192280
#> 98  31.07819 4.192280
#> 99  31.16748 4.192280
#> 100 30.09438 4.075283
#> 101 30.65169 4.075283
#> 102 30.09438 4.075283
#> 103 30.56240 4.075283
#> 104 30.09438 4.075283
#> 105 30.56240 4.075283
#> 106 30.47311 4.075283
#> 107 29.88638 4.075283
#> 108 30.32075 4.192280
#> 109 33.02060 4.312635
#> 110 33.44083 4.192280
#> 111 33.06211 4.192280
#> 112 19.65056 1.952332
#> 113 19.29442 1.952332
#> 114 19.08641 1.952332
#> 115 18.73027 1.952332
#> 116 19.16099 1.952332
#> 117 19.00499 1.952332
#> 118 19.27913 2.066039
#> 119 19.38735 2.008381
#> 120 18.73027 1.952332
#> 121 21.44905 1.952332
#> 122 21.76106 1.952332
#> 123 22.45179 1.952332
#> 124 21.58248 1.952332
#> 125 21.80884 2.008381