Summarising Smooth Information Criterion (SIC) Fits

summary method class “smoothic”

Usage

# S3 method for class 'smoothic'
summary(object, ...)

Arguments

object

an object of class “smoothic” which is the result of a call to smoothic.

...

further arguments passed to or from other methods.

Value

A list containing the following components:

model - the matched model from the smoothic object.

coefmat - a typical coefficient matrix whose columns are the estimated regression coefficients, estimated standard errors (SEE) and p-values.

plike - value of the penalized likelihood function.

Author

Meadhbh O'Neill

Examples

# Sniffer Data --------------------
# MPR Model ----
results <- smoothic(
  formula = y ~ .,
  data = sniffer,
  family = "normal",
  model = "mpr"
)
summary(results)
#> Call:
#> smoothic(formula = y ~ ., data = sniffer, family = "normal", 
#>     model = "mpr")
#> Family:
#> [1] "normal"
#> Model:
#> [1] "mpr"
#> 
#> Coefficients:
#> 
#> Location:
#>                   Estimate        SE       Z    Pvalue    
#> intercept_0_beta  0.740133  0.922060  0.8027  0.177285    
#> tanktemp_1_beta  -0.089276  0.040426 -2.2084  0.001239 ** 
#> gastemp_2_beta    0.226337  0.028123  8.0481 < 2.2e-16 ***
#> tankpres_3_beta          0         0       0         0    
#> gaspres_4_beta    5.200004  0.837714  6.2074 < 2.2e-16 ***
#> 
#> Scale:
#>                    Estimate        SE       Z    Pvalue    
#> intercept_0_alpha -0.646010  0.724506 -0.8917    0.1434    
#> tanktemp_1_alpha          0         0       0         0    
#> gastemp_2_alpha    0.056654  0.011278  5.0233 8.382e-13 ***
#> tankpres_3_alpha          0         0       0         0    
#> gaspres_4_alpha           0         0       0         0    
#> 
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#> Penalized Likelihood:
#> [1] -310.633
#> IC Value:
#> [1] 621.2659