Predict smoothic

predict method class “smoothic”

Usage

# S3 method for smoothic
predict(object, newdata, ...)

Arguments

object

an object of class “smoothic” which is the result of a call to smoothic.

newdata

new data object

...

further arguments passed to or from other methods.

Value

a matrix containing the predicted values for the location mu and scale s

Author

Meadhbh O'Neill

Examples

# Sniffer Data --------------------
# MPR Model ----
results <- smoothic(
  formula = y ~ .,
  data = sniffer,
  family = "normal",
  model = "mpr"
)
predict(results)
#>           mu        s
#> 1   23.85783 1.846919
#> 2   26.20666 2.821905
#> 3   27.57491 3.250179
#> 4   26.97760 3.159616
#> 5   26.36969 2.520290
#> 6   23.07232 2.010327
#> 7   22.00986 1.899856
#> 8   22.36969 1.954311
#> 9   25.87344 3.071576
#> 10  24.81460 2.985990
#> 11  25.21177 3.071576
#> 12  24.99569 3.071576
#> 13  27.58523 3.343337
#> 14  27.60789 3.343337
#> 15  27.48118 3.343337
#> 16  28.30850 3.639143
#> 17  21.45150 2.010327
#> 18  21.29180 1.954311
#> 19  20.59354 1.846919
#> 20  20.96496 1.954311
#> 21  32.01770 3.537742
#> 22  32.50424 3.639143
#> 23  33.28390 3.850748
#> 24  33.51031 3.961120
#> 25  33.51031 3.961120
#> 26  33.51031 3.961120
#> 27  31.98683 3.961120
#> 28  31.69733 3.961120
#> 29  31.47092 3.850748
#> 30  32.52179 3.961120
#> 31  32.43242 3.961120
#> 32  32.34305 3.961120
#> 33  31.98322 3.850748
#> 34  32.20963 3.961120
#> 35  31.62264 3.961120
#> 36  31.87916 3.850748
#> 37  31.52966 3.850748
#> 38  30.59753 3.743450
#> 39  31.41092 3.850748
#> 40  31.63733 3.961120
#> 41  32.17589 4.074657
#> 42  32.40230 4.191447
#> 43  31.60359 4.074657
#> 44  31.25845 3.961120
#> 45  24.01845 2.010327
#> 46  23.37944 2.010327
#> 47  25.59681 2.188193
#> 48  21.95129 2.127221
#> 49  20.58901 1.954311
#> 50  30.08833 3.639143
#> 51  30.73818 3.961120
#> 52  30.75286 3.961120
#> 53  31.30612 4.074657
#> 54  31.84107 4.074657
#> 55  31.39549 4.074657
#> 56  30.97927 4.074657
#> 57  42.03623 4.435166
#> 58  41.23317 4.435166
#> 59  42.82635 3.537742
#> 60  43.73197 3.961120
#> 61  51.00346 8.989474
#> 62  52.09889 9.784828
#> 63  52.09889 9.784828
#> 64  52.09889 9.784828
#> 65  50.65756 9.247136
#> 66  51.19975 9.784828
#> 67  32.54145 4.965944
#> 68  29.98639 4.191447
#> 69  30.54400 4.191447
#> 70  30.07576 4.191447
#> 71  23.83797 2.520290
#> 72  22.91692 2.381796
#> 73  41.00887 4.965944
#> 74  39.96908 4.693057
#> 75  37.75893 4.562290
#> 76  39.02077 5.108281
#> 77  45.92022 8.495488
#> 78  47.08673 8.989474
#> 79  47.40251 9.247136
#> 80  31.16833 4.191447
#> 81  31.50986 4.191447
#> 82  31.03129 4.074657
#> 83  31.07896 4.191447
#> 84  30.51102 4.074657
#> 85  30.18417 4.074657
#> 86  30.56305 4.074657
#> 87  30.56305 4.074657
#> 88  30.65242 4.074657
#> 89  30.78946 4.191447
#> 90  30.11310 4.191447
#> 91  30.92649 4.311585
#> 92  31.34707 4.191447
#> 93  32.09375 4.311585
#> 94  31.16833 4.191447
#> 95  30.70009 4.191447
#> 96  30.41058 4.191447
#> 97  30.20247 4.191447
#> 98  31.07896 4.191447
#> 99  31.16833 4.191447
#> 100 30.09480 4.074657
#> 101 30.65242 4.074657
#> 102 30.09480 4.074657
#> 103 30.56305 4.074657
#> 104 30.09480 4.074657
#> 105 30.56305 4.074657
#> 106 30.47368 4.074657
#> 107 29.88670 4.074657
#> 108 30.32121 4.191447
#> 109 33.02227 4.311585
#> 110 33.44284 4.191447
#> 111 33.06397 4.191447
#> 112 19.64833 1.954311
#> 113 19.29211 1.954311
#> 114 19.08400 1.954311
#> 115 18.72778 1.954311
#> 116 19.15869 1.954311
#> 117 19.00261 1.954311
#> 118 19.27668 2.067948
#> 119 19.38509 2.010327
#> 120 18.72778 1.954311
#> 121 21.44789 1.954311
#> 122 21.76005 1.954311
#> 123 22.45109 1.954311
#> 124 21.58131 1.954311
#> 125 21.80772 2.010327