Predict smoothic

predict method class “smoothic”

Usage

# S3 method for smoothic
predict(object, newdata, ...)

Arguments

object

an object of class “smoothic” which is the result of a call to smoothic.

newdata

new data object

...

further arguments passed to or from other methods.

Value

a matrix containing the predicted values for the location mu and scale s

Author

Meadhbh O'Neill

Examples

# Sniffer Data --------------------
# MPR Model ----
results <- smoothic(
  formula = y ~ .,
  data = sniffer,
  family = "normal",
  model = "mpr"
)
predict(results)
#>           mu        s
#> 1   23.73024 1.010112
#> 2   26.07308 2.020696
#> 3   27.52946 3.135873
#> 4   26.88822 2.671811
#> 5   26.31907 1.798780
#> 6   22.92405 1.565104
#> 7   21.83871 1.682448
#> 8   22.22059 1.729886
#> 9   25.77555 3.299793
#> 10  24.68304 3.555869
#> 11  25.09690 3.755235
#> 12  24.89318 4.347321
#> 13  27.55603 3.393922
#> 14  27.56611 2.943488
#> 15  27.44825 2.986425
#> 16  28.31262 3.567249
#> 17  21.26532 1.881589
#> 18  21.11083 1.797198
#> 19  20.40271 1.768715
#> 20  20.76559 1.752554
#> 21  32.29267 3.548556
#> 22  32.79240 3.508289
#> 23  33.59822 3.558374
#> 24  33.82851 3.466691
#> 25  33.82851 3.466691
#> 26  33.82851 3.466691
#> 27  32.24383 4.315213
#> 28  31.93057 3.756791
#> 29  31.70028 3.609989
#> 30  32.80462 3.601584
#> 31  32.71875 3.943520
#> 32  32.63288 4.282463
#> 33  32.25100 3.650248
#> 34  32.48129 3.798688
#> 35  31.86661 4.165089
#> 36  32.14322 3.516922
#> 37  31.78791 3.857758
#> 38  30.81325 3.931877
#> 39  31.65822 3.701174
#> 40  31.88851 3.851684
#> 41  32.44214 3.499551
#> 42  32.67243 3.641862
#> 43  31.84936 3.917426
#> 44  31.48939 3.552484
#> 45  24.09821 2.044922
#> 46  23.42964 2.228225
#> 47  25.72710 2.093945
#> 48  21.96205 3.943120
#> 49  20.53689 3.326342
#> 50  30.30344 3.731401
#> 51  30.95050 3.871732
#> 52  30.97240 3.856205
#> 53  31.54793 3.949035
#> 54  32.10873 3.854965
#> 55  31.63380 3.696948
#> 56  31.20269 4.013020
#> 57  42.84352 4.554228
#> 58  42.02511 6.488188
#> 59  43.69563 1.982669
#> 60  44.61680 2.325386
#> 61  52.03959 6.908159
#> 62  53.16157 7.172456
#> 63  53.16157 7.172456
#> 64  53.16157 7.172456
#> 65  51.67710 7.596182
#> 66  52.22355 7.451545
#> 67  32.80465 4.361844
#> 68  30.18346 4.347439
#> 69  30.75434 3.711114
#> 70  30.26934 4.069920
#> 71  23.87085 2.761193
#> 72  22.92527 2.796120
#> 73  41.78262 6.034589
#> 74  40.70736 5.672669
#> 75  38.39555 5.966792
#> 76  39.70577 7.418645
#> 77  46.78290 9.296929
#> 78  47.97602 8.198046
#> 79  48.29218 7.856191
#> 80  31.40100 3.837960
#> 81  31.76815 4.084929
#> 82  31.25658 3.452563
#> 83  31.31513 4.167830
#> 84  30.71769 3.825398
#> 85  30.37246 4.017731
#> 86  30.77158 3.786371
#> 87  30.77158 3.786371
#> 88  30.85745 3.663526
#> 89  31.00187 3.940346
#> 90  30.30132 4.502252
#> 91  31.14629 4.416458
#> 92  31.57275 3.476394
#> 93  32.34192 3.265166
#> 94  31.40100 3.869737
#> 95  30.91600 4.243878
#> 96  30.60275 4.181114
#> 97  30.38719 4.356182
#> 98  31.31513 4.167830
#> 99  31.40100 3.869737
#> 100 30.28658 4.152452
#> 101 30.85745 3.663526
#> 102 30.28658 4.152452
#> 103 30.77158 3.786371
#> 104 30.28658 4.152452
#> 105 30.77158 3.786371
#> 106 30.68571 4.078042
#> 107 30.07103 4.326320
#> 108 30.51687 4.321314
#> 109 33.32375 4.371754
#> 110 33.75021 3.412948
#> 111 33.35108 3.475222
#> 112 19.44203 2.183341
#> 113 19.05297 2.128661
#> 114 18.83742 2.093408
#> 115 18.44837 2.162247
#> 116 18.90138 1.903827
#> 117 18.73972 1.963306
#> 118 19.02856 2.386402
#> 119 19.13168 1.981247
#> 120 18.44837 2.162247
#> 121 21.27249 2.055202
#> 122 21.59583 1.932562
#> 123 22.31829 1.712595
#> 124 21.42408 1.980963
#> 125 21.65437 2.061520