Predict smoothic

predict method class “smoothic”

Usage

# S3 method for smoothic
predict(object, newdata, ...)

Arguments

object

an object of class “smoothic” which is the result of a call to smoothic.

newdata

new data object

...

further arguments passed to or from other methods.

Value

a matrix containing the predicted values for the location mu and scale s

Author

Meadhbh O'Neill

Examples

# Sniffer Data --------------------
# MPR Model ----
results <- smoothic(
  formula = y ~ .,
  data = sniffer,
  family = "normal",
  model = "mpr"
)
predict(results)
#>           mu        s
#> 1   23.85753 1.843722
#> 2   26.20568 2.819873
#> 3   27.57388 3.248934
#> 4   26.97665 3.158192
#> 5   26.36885 2.517795
#> 6   23.07271 2.007254
#> 7   22.01076 1.896696
#> 8   22.37055 1.951192
#> 9   25.87321 3.069985
#> 10  24.81487 2.984242
#> 11  25.21193 3.069985
#> 12  24.99610 3.069985
#> 13  27.58439 3.342282
#> 14  27.60694 3.342282
#> 15  27.48039 3.342282
#> 16  28.30757 3.638731
#> 17  21.45288 2.007254
#> 18  21.29327 1.951192
#> 19  20.59532 1.843722
#> 20  20.96654 1.951192
#> 21  32.01572 3.537102
#> 22  32.50206 3.638731
#> 23  33.28146 3.850831
#> 24  33.50779 3.961473
#> 25  33.50779 3.961473
#> 26  33.50779 3.961473
#> 27  31.98507 3.961473
#> 28  31.69562 3.961473
#> 29  31.46928 3.850831
#> 30  32.51979 3.961473
#> 31  32.43052 3.961473
#> 32  32.34124 3.961473
#> 33  31.98146 3.850831
#> 34  32.20779 3.961473
#> 35  31.62107 3.961473
#> 36  31.87746 3.850831
#> 37  31.52818 3.850831
#> 38  30.59639 3.743279
#> 39  31.40946 3.850831
#> 40  31.63579 3.961473
#> 41  32.17413 4.075294
#> 42  32.40047 4.192386
#> 43  31.60213 4.075294
#> 44  31.25707 3.961473
#> 45  24.01971 2.007254
#> 46  23.38099 2.007254
#> 47  25.59745 2.185290
#> 48  21.95383 2.124256
#> 49  20.59193 1.951192
#> 50  30.08751 3.638731
#> 51  30.73707 3.961473
#> 52  30.75179 3.961473
#> 53  31.30485 4.075294
#> 54  31.83958 4.075294
#> 55  31.39413 4.075294
#> 56  30.97813 4.075294
#> 57  42.03087 4.436758
#> 58  41.22832 4.436758
#> 59  42.82018 3.537102
#> 60  43.72553 3.961473
#> 61  50.99474 9.007877
#> 62  52.08975 9.806844
#> 63  52.08975 9.806844
#> 64  52.08975 9.806844
#> 65  50.64907 9.266692
#> 66  51.19102 9.806844
#> 67  32.53993 4.969067
#> 68  29.98591 4.192386
#> 69  30.54319 4.192386
#> 70  30.07519 4.192386
#> 71  23.83968 2.517795
#> 72  22.91901 2.379118
#> 73  41.00439 4.969067
#> 74  39.96499 4.695375
#> 75  37.75576 4.564236
#> 76  39.01728 5.111838
#> 77  45.91406 8.511732
#> 78  47.08001 9.007877
#> 79  47.39562 9.266692
#> 80  31.16719 4.192386
#> 81  31.50864 4.192386
#> 82  31.03013 4.075294
#> 83  31.07791 4.192386
#> 84  30.51013 4.075294
#> 85  30.18340 4.075294
#> 86  30.56213 4.075294
#> 87  30.56213 4.075294
#> 88  30.65140 4.075294
#> 89  30.78846 4.192386
#> 90  30.11246 4.192386
#> 91  30.92553 4.312842
#> 92  31.34574 4.192386
#> 93  32.09208 4.312842
#> 94  31.16719 4.192386
#> 95  30.69919 4.192386
#> 96  30.40974 4.192386
#> 97  30.20174 4.192386
#> 98  31.07791 4.192386
#> 99  31.16719 4.192386
#> 100 30.09413 4.075294
#> 101 30.65140 4.075294
#> 102 30.09413 4.075294
#> 103 30.56213 4.075294
#> 104 30.09413 4.075294
#> 105 30.56213 4.075294
#> 106 30.47285 4.075294
#> 107 29.88613 4.075294
#> 108 30.32046 4.192386
#> 109 33.02025 4.312842
#> 110 33.44047 4.192386
#> 111 33.06174 4.192386
#> 112 19.65089 1.951192
#> 113 19.29472 1.951192
#> 114 19.08672 1.951192
#> 115 18.73054 1.951192
#> 116 19.16127 1.951192
#> 117 19.00527 1.951192
#> 118 19.27939 2.064926
#> 119 19.38760 2.007254
#> 120 18.73054 1.951192
#> 121 21.44927 1.951192
#> 122 21.76127 1.951192
#> 123 22.45199 1.951192
#> 124 21.58272 1.951192
#> 125 21.80905 2.007254