Predict smoothic

predict method class “smoothic”

Usage

# S3 method for smoothic
predict(object, newdata, ...)

Arguments

object

an object of class “smoothic” which is the result of a call to smoothic.

newdata

new data object

...

further arguments passed to or from other methods.

Value

a matrix containing the predicted values for the location mu and scale s

Author

Meadhbh O'Neill

Examples

# Sniffer Data --------------------
# MPR Model ----
results <- smoothic(
  formula = y ~ .,
  data = sniffer,
  family = "normal",
  model = "mpr"
)
predict(results)
#>           mu        s
#> 1   23.85226 1.841728
#> 2   26.19990 2.818753
#> 3   27.56963 3.248386
#> 4   26.97209 3.157515
#> 5   26.36388 2.516336
#> 6   23.07070 2.005357
#> 7   22.01034 1.894731
#> 8   22.37033 1.949259
#> 9   25.87146 3.069187
#> 10  24.81476 2.983330
#> 11  25.21168 3.069187
#> 12  24.99707 3.069187
#> 13  27.58125 3.341871
#> 14  27.60322 3.341871
#> 15  27.47745 3.341871
#> 16  28.30489 3.638782
#> 17  21.45470 2.005357
#> 18  21.29532 1.949259
#> 19  20.59829 1.841728
#> 20  20.96894 1.949259
#> 21  32.01120 3.536991
#> 22  32.49696 3.638782
#> 23  33.27584 3.851237
#> 24  33.50208 3.962072
#> 25  33.50208 3.962072
#> 26  33.50208 3.962072
#> 27  31.98192 3.962072
#> 28  31.69247 3.962072
#> 29  31.46623 3.851237
#> 30  32.51592 3.962072
#> 31  32.42708 3.962072
#> 32  32.33824 3.962072
#> 33  31.97826 3.851237
#> 34  32.20450 3.962072
#> 35  31.61860 3.962072
#> 36  31.87445 3.851237
#> 37  31.52610 3.851237
#> 38  30.59518 3.743503
#> 39  31.40732 3.851237
#> 40  31.63357 3.962072
#> 41  32.17123 4.076097
#> 42  32.39747 4.193403
#> 43  31.60030 4.076097
#> 44  31.25528 3.962072
#> 45  24.02377 2.005357
#> 46  23.38597 2.005357
#> 47  25.59983 2.183524
#> 48  21.96294 2.122443
#> 49  20.60150 1.949259
#> 50  30.08745 3.638782
#> 51  30.73625 3.962072
#> 52  30.75121 3.962072
#> 53  31.30384 4.076097
#> 54  31.83784 4.076097
#> 55  31.39268 4.076097
#> 56  30.97746 4.076097
#> 57  42.01871 4.438240
#> 58  41.21819 4.438240
#> 59  42.80392 3.536991
#> 60  43.70889 3.962072
#> 61  50.97465 9.021183
#> 62  52.06861 9.822676
#> 63  52.06861 9.822676
#> 64  52.06861 9.822676
#> 65  50.62996 9.280804
#> 66  51.17129 9.822676
#> 67  32.53920 4.971635
#> 68  29.98796 4.193403
#> 69  30.54393 4.193403
#> 70  30.07680 4.193403
#> 71  23.84644 2.516336
#> 72  22.92682 2.377521
#> 73  40.99626 4.971635
#> 74  39.95787 4.697373
#> 75  37.75134 4.565968
#> 76  39.01264 5.114714
#> 77  45.90278 8.523526
#> 78  47.06694 9.021183
#> 79  47.38203 9.280804
#> 80  31.16676 4.193403
#> 81  31.50812 4.193403
#> 82  31.02936 4.076097
#> 83  31.07792 4.193403
#> 84  30.51033 4.076097
#> 85  30.18394 4.076097
#> 86  30.56223 4.076097
#> 87  30.56223 4.076097
#> 88  30.65107 4.076097
#> 89  30.78847 4.193403
#> 90  30.11373 4.193403
#> 91  30.92588 4.314085
#> 92  31.34444 4.193403
#> 93  32.08972 4.314085
#> 94  31.16676 4.193403
#> 95  30.69964 4.193403
#> 96  30.41019 4.193403
#> 97  30.20257 4.193403
#> 98  31.07792 4.193403
#> 99  31.16676 4.193403
#> 100 30.09510 4.076097
#> 101 30.65107 4.076097
#> 102 30.09510 4.076097
#> 103 30.56223 4.076097
#> 104 30.09510 4.076097
#> 105 30.56223 4.076097
#> 106 30.47339 4.076097
#> 107 29.88749 4.076097
#> 108 30.32135 4.193403
#> 109 33.01698 4.314085
#> 110 33.43554 4.193403
#> 111 33.05725 4.193403
#> 112 19.65735 1.949259
#> 113 19.30103 1.949259
#> 114 19.09342 1.949259
#> 115 18.73710 1.949259
#> 116 19.16729 1.949259
#> 117 19.01158 1.949259
#> 118 19.28639 2.063069
#> 119 19.39354 2.005357
#> 120 18.73710 1.949259
#> 121 21.45103 1.949259
#> 122 21.76245 1.949259
#> 123 22.45216 1.949259
#> 124 21.58477 1.949259
#> 125 21.81102 2.005357