Predict smoothic

predict method class “smoothic”

Usage

# S3 method for smoothic
predict(object, newdata, ...)

Arguments

object

an object of class “smoothic” which is the result of a call to smoothic.

newdata

new data object

...

further arguments passed to or from other methods.

Value

a matrix containing the predicted values for the location mu and scale s

Author

Meadhbh O'Neill

Examples

# Sniffer Data --------------------
# MPR Model ----
results <- smoothic(
  formula = y ~ .,
  data = sniffer,
  family = "normal",
  model = "mpr"
)
predict(results)
#>           mu        s
#> 1   23.76345 1.060816
#> 2   26.10793 2.071384
#> 3   27.55003 3.139367
#> 4   26.91465 2.702051
#> 5   26.34235 1.846596
#> 6   22.95108 1.598818
#> 7   21.86429 1.702663
#> 8   22.24308 1.750391
#> 9   25.79568 3.277637
#> 10  24.70280 3.504933
#> 11  25.11508 3.693324
#> 12  24.90677 4.228457
#> 13  27.57178 3.383851
#> 14  27.58482 2.966560
#> 15  27.46392 3.006551
#> 16  28.32327 3.563724
#> 17  21.28675 1.895099
#> 18  21.13106 1.812994
#> 19  20.42176 1.779589
#> 20  20.78706 1.771232
#> 21  32.27688 3.542525
#> 22  32.77657 3.512174
#> 23  33.58068 3.572284
#> 24  33.81074 3.493794
#> 25  33.81074 3.493794
#> 26  33.81074 3.493794
#> 27  32.22643 4.276358
#> 28  31.91591 3.762063
#> 29  31.68586 3.619097
#> 30  32.78613 3.618749
#> 31  32.69872 3.935147
#> 32  32.61131 4.246725
#> 33  32.23252 3.656777
#> 34  32.46258 3.801232
#> 35  31.84894 4.138479
#> 36  32.12467 3.533120
#> 37  31.76762 3.848289
#> 38  30.79564 3.908630
#> 39  31.63933 3.703605
#> 40  31.86938 3.849909
#> 41  32.42300 3.530345
#> 42  32.65306 3.669805
#> 43  31.82981 3.917915
#> 44  31.47146 3.572462
#> 45  24.08313 2.048031
#> 46  23.41557 2.216904
#> 47  25.71049 2.105657
#> 48  21.94172 3.770360
#> 49  20.52267 3.203077
#> 50  30.28291 3.716949
#> 51  30.93219 3.867892
#> 52  30.95263 3.853611
#> 53  31.52669 3.947007
#> 54  32.08640 3.860373
#> 55  31.61410 3.713585
#> 56  31.18269 4.005841
#> 57  42.78604 4.536706
#> 58  41.96410 6.291468
#> 59  43.64192 2.072561
#> 60  44.56214 2.419973
#> 61  51.97569 6.996274
#> 62  53.09727 7.285241
#> 63  53.09727 7.285241
#> 64  53.09727 7.285241
#> 65  51.61256 7.651830
#> 66  52.16008 7.545797
#> 67  32.78004 4.384806
#> 68  30.15939 4.320996
#> 69  30.73214 3.733331
#> 70  30.24680 4.065457
#> 71  23.85393 2.743848
#> 72  22.90848 2.765166
#> 73  41.71838 5.928053
#> 74  40.64464 5.576902
#> 75  38.33754 5.830585
#> 76  39.64264 7.185713
#> 77  46.71613 9.164851
#> 78  47.91077 8.190933
#> 79  48.22824 7.889800
#> 80  31.37926 3.851440
#> 81  31.74370 4.080225
#> 82  31.23662 3.485963
#> 83  31.29185 4.156390
#> 84  30.69735 3.832200
#> 85  30.35335 4.009693
#> 86  30.75128 3.796082
#> 87  30.75128 3.796082
#> 88  30.83869 3.682123
#> 89  30.98133 3.946039
#> 90  30.28029 4.462974
#> 91  31.12398 4.393100
#> 92  31.55408 3.514886
#> 93  32.32341 3.323623
#> 94  31.37926 3.880900
#> 95  30.89392 4.226153
#> 96  30.58341 4.168089
#> 97  30.36770 4.328995
#> 98  31.29185 4.156390
#> 99  31.37926 3.880900
#> 100 30.26594 4.133790
#> 101 30.83869 3.682123
#> 102 30.26594 4.133790
#> 103 30.75128 3.796082
#> 104 30.26594 4.133790
#> 105 30.75128 3.796082
#> 106 30.66387 4.065551
#> 107 30.05023 4.293373
#> 108 30.49600 4.297088
#> 109 33.30148 4.353311
#> 110 33.73159 3.456612
#> 111 33.33366 3.514630
#> 112 19.45369 2.169892
#> 113 19.06881 2.119439
#> 114 18.85311 2.086920
#> 115 18.46822 2.150076
#> 116 18.92008 1.911615
#> 117 18.75830 1.966700
#> 118 19.04359 2.364395
#> 119 19.15013 1.987130
#> 120 18.46822 2.150076
#> 121 21.29284 2.052320
#> 122 21.61640 1.938964
#> 123 22.33789 1.734279
#> 124 21.44158 1.983800
#> 125 21.67163 2.062166