Predict smoothic

predict method class “smoothic”

Usage

# S3 method for smoothic
predict(object, newdata, ...)

Arguments

object

an object of class “smoothic” which is the result of a call to smoothic.

newdata

new data object

...

further arguments passed to or from other methods.

Value

a matrix containing the predicted values for the location mu and scale s

Author

Meadhbh O'Neill

Examples

# Sniffer Data --------------------
# MPR Model ----
results <- smoothic(
  formula = y ~ .,
  data = sniffer,
  family = "normal",
  model = "mpr"
)
predict(results)
#>           mu        s
#> 1   23.85760 1.843143
#> 2   26.20577 2.819553
#> 3   27.57393 3.248783
#> 4   26.97674 3.158004
#> 5   26.36887 2.517375
#> 6   23.07293 2.006704
#> 7   22.01106 1.896125
#> 8   22.37082 1.950631
#> 9   25.87340 3.069761
#> 10  24.81516 2.983984
#> 11  25.21220 3.069761
#> 12  24.99640 3.069761
#> 13  27.58446 3.342172
#> 14  27.60701 3.342172
#> 15  27.48047 3.342172
#> 16  28.30763 3.638758
#> 17  21.45328 2.006704
#> 18  21.29366 1.950631
#> 19  20.59575 1.843143
#> 20  20.96697 1.950631
#> 21  32.01540 3.537081
#> 22  32.50170 3.638758
#> 23  33.28106 3.850963
#> 24  33.50739 3.961662
#> 25  33.50739 3.961662
#> 26  33.50739 3.961662
#> 27  31.98482 3.961662
#> 28  31.69540 3.961662
#> 29  31.46907 3.850963
#> 30  32.51949 3.961662
#> 31  32.43023 3.961662
#> 32  32.34096 3.961662
#> 33  31.98119 3.850963
#> 34  32.20752 3.961662
#> 35  31.62086 3.961662
#> 36  31.87720 3.850963
#> 37  31.52797 3.850963
#> 38  30.59626 3.743357
#> 39  31.40925 3.850963
#> 40  31.63558 3.961662
#> 41  32.17388 4.075543
#> 42  32.40021 4.192698
#> 43  31.60194 4.075543
#> 44  31.25690 3.961662
#> 45  24.01988 2.006704
#> 46  23.38122 2.006704
#> 47  25.59750 2.184779
#> 48  21.95426 2.123731
#> 49  20.59243 1.950631
#> 50  30.08741 3.638758
#> 51  30.73695 3.961662
#> 52  30.75168 3.961662
#> 53  31.30470 4.075543
#> 54  31.83937 4.075543
#> 55  31.39396 4.075543
#> 56  30.97801 4.075543
#> 57  42.02968 4.437208
#> 58  41.22722 4.437208
#> 59  42.81876 3.537081
#> 60  43.72409 3.961662
#> 61  50.99305 9.011810
#> 62  52.08800 9.811520
#> 63  52.08800 9.811520
#> 64  52.08800 9.811520
#> 65  50.64744 9.270862
#> 66  51.18937 9.811520
#> 67  32.53978 4.969837
#> 68  29.98592 4.192698
#> 69  30.54313 4.192698
#> 70  30.07518 4.192698
#> 71  23.84001 2.517375
#> 72  22.91940 2.378657
#> 73  41.00339 4.969837
#> 74  39.96406 4.695977
#> 75  37.75504 4.564759
#> 76  39.01650 5.112699
#> 77  45.91288 8.515219
#> 78  47.07874 9.011810
#> 79  47.39434 9.270862
#> 80  31.16707 4.192698
#> 81  31.50848 4.192698
#> 82  31.03000 4.075543
#> 83  31.07780 4.192698
#> 84  30.51006 4.075543
#> 85  30.18337 4.075543
#> 86  30.56205 4.075543
#> 87  30.56205 4.075543
#> 88  30.65132 4.075543
#> 89  30.78838 4.192698
#> 90  30.11245 4.192698
#> 91  30.92545 4.313221
#> 92  31.34560 4.192698
#> 93  32.09187 4.313221
#> 94  31.16707 4.192698
#> 95  30.69912 4.192698
#> 96  30.40970 4.192698
#> 97  30.20172 4.192698
#> 98  31.07780 4.192698
#> 99  31.16707 4.192698
#> 100 30.09410 4.075543
#> 101 30.65132 4.075543
#> 102 30.09410 4.075543
#> 103 30.56205 4.075543
#> 104 30.09410 4.075543
#> 105 30.56205 4.075543
#> 106 30.47279 4.075543
#> 107 29.88612 4.075543
#> 108 30.32043 4.192698
#> 109 33.01995 4.313221
#> 110 33.44010 4.192698
#> 111 33.06142 4.192698
#> 112 19.65147 1.950631
#> 113 19.29533 1.950631
#> 114 19.08735 1.950631
#> 115 18.73121 1.950631
#> 116 19.16189 1.950631
#> 117 19.00591 1.950631
#> 118 19.28004 2.064388
#> 119 19.38822 2.006704
#> 120 18.73121 1.950631
#> 121 21.44965 1.950631
#> 122 21.76161 1.950631
#> 123 22.45227 1.950631
#> 124 21.58309 1.950631
#> 125 21.80942 2.006704