Variable Selection Using a Smooth Information Criterion (SIC)

Implements the SIC \(\epsilon\)-telescope method, either using single or multi-parameter regression. Returns estimated coefficients, estimated standard errors (SEE) and the value of the penalized likelihood function. Note that the function will scale the predictors to have unit variance, however, the final estimates are converted back to their original scale.

Usage

smoothic(
  formula,
  data,
  family = "sgnd",
  model = "mpr",
  lambda = "log(n)",
  epsilon_1 = 0.3,
  epsilon_T = 1e-04,
  steps_T = 100,
  zero_tol = 1e-05,
  max_it = 10000,
  optimizer = "nlm",
  kappa,
  tau,
  stepmax_nlm
)

Arguments

formula

An object of class "formula": a two-sided object with response on the left hand side and the model variables on the right hand side.

data

A data frame containing the variables in the model; the data frame should be unstandardized.

family

The family of the model, default is family = "sgnd" for the "Smooth Generalized Distribution" where the shape parameter kappa is also estimated. Classical regression with normally distributed errors is performed when family = "normal". If family = "laplace", this corresponds to a robust regression with errors from the Laplace distribution.

model

The type of regression to be implemented, either model = "mpr" for multi-parameter regression, or model = "spr" for single parameter regression (i.e., classical normal linear regression). Defaults to model="mpr".

lambda

Value of penalty tuning parameter. Suggested values are "log(n)" and "2" for the BIC and AIC respectively. Defaults to lambda ="log(n)" for the BIC case.

epsilon_1

Starting value for \(\epsilon\)-telescope. Defaults to 0.3.

epsilon_T

Final value for \(\epsilon\)-telescope. Defaults to 1e-04.

steps_T

Number of steps in \(\epsilon\)-telescope. Defaults to 100.

zero_tol

Coefficients below this value are treated as being zero. Defaults to 1e-05.

max_it

Maximum number of iterations to performed before the optimization is terminated. Defaults to 1e+04.

optimizer

The optimization procedure to be used. Defaults to optimizer = "nlm", where the nlm function from the stats package is used. This tends to be more stable than the manually coded Newton-Raphson procedure that is used when optimizer = "manual".

kappa

Optional user-supplied positive kappa value (> 0.2 to avoid computational issues) if family = "sgnd". If supplied, the shape parameter kappa will be fixed to this value in the optimization. If not supplied, kappa is estimated from the data.

tau

Optional user-supplied positive smoothing parameter value in the "Smooth Generalized Normal Distribution" if family = "sgnd" or family = "laplace". If not supplied, then tau = "0.15". Smaller values of tau bring the approximation closer to the absolute value function, but this can cause the optimization to become unstable. Some issues with standard error calculation with smaller values of tau when using the Laplace distribution in the robust regression setting.

stepmax_nlm

Optional maximum allowable scaled step length (positive scalar) to be passed to nlm if optimizer = "nlm". If not supplied, default values in nlm are used.

Value

A list with estimates and estimated standard errors.

coefficients - vector of coefficients.

see - vector of estimated standard errors.

model - the matched type of model which is called.

plike - value of the penalized likelihood function.

kappa - value of the estimated/fixed shape parameter kappa if family = "sgnd".

References

O'Neill, M. and Burke, K. (2021) Variable Selection Using a Smooth Information Criterion for Multi-Parameter Regression Models. <arXiv:2110.02643>

O'Neill, M. and Burke, K. (2022) Robust Distributional Regression with Automatic Variable Selection. <soon to be on arXiv>

Author

Meadhbh O'Neill

Examples

# Sniffer Data --------------------
# MPR Model ----
results <- smoothic(
  formula = y ~ .,
  data = sniffer,
  family = "normal",
  model = "mpr"
)
summary(results)
#> Call:
#> smoothic(formula = y ~ ., data = sniffer, family = "normal", 
#>     model = "mpr")
#> Family:
#> [1] "normal"
#> Model:
#> [1] "mpr"
#> 
#> Coefficients:
#>                    Estimate       SEE       Z    Pvalue    
#> intercept_0_beta   0.758849  0.853924  0.8887   0.36670    
#> tanktemp_1_beta   -0.088858  0.027623 -3.2168   0.00127 ** 
#> gastemp_2_beta     0.226261  0.025302  8.9423 < 2.2e-16 ***
#> tankpres_3_beta           0         0       0         0    
#> gaspres_4_beta     5.190518  0.510752 10.1625 < 2.2e-16 ***
#> intercept_0_alpha -1.343862  0.643686 -2.0878   0.03608 *  
#> tanktemp_1_alpha          0         0       0         0    
#> gastemp_2_alpha    0.056737  0.011170  5.0792 3.714e-07 ***
#> tankpres_3_alpha          0         0       0         0    
#> gaspres_4_alpha           0         0       0         0    
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> Penalized Likelihood:
#> [1] -308.2184