Summarising Smooth Information Criterion (SIC) Fits

summary method class “smoothic”

Usage

# S3 method for smoothic
summary(object, ...)

Arguments

object

an object of class “smoothic” which is the result of a call to smoothic.

...

further arguments passed to or from other methods.

Value

A list containing the following components:

model - the matched model from the smoothic object.

coefmat - a typical coefficient matrix whose columns are the estimated regression coefficients, estimated standard errors (SEE) and p-values.

plike - value of the penalized likelihood function.

Author

Meadhbh O'Neill

Examples

# Sniffer Data --------------------
# MPR Model ----
results <- smoothic(
  formula = y ~ .,
  data = sniffer,
  family = "normal",
  model = "mpr"
)
summary(results)
#> Call:
#> smoothic(formula = y ~ ., data = sniffer, family = "normal", 
#>     model = "mpr")
#> Family:
#> [1] "normal"
#> Model:
#> [1] "mpr"
#> 
#> Coefficients:
#>                    Estimate       SEE       Z    Pvalue    
#> intercept_0_beta   0.758850  0.853924  0.8887   0.36670    
#> tanktemp_1_beta   -0.088858  0.027623 -3.2168   0.00127 ** 
#> gastemp_2_beta     0.226261  0.025302  8.9423 < 2.2e-16 ***
#> tankpres_3_beta           0         0       0         0    
#> gaspres_4_beta     5.190518  0.510752 10.1625 < 2.2e-16 ***
#> intercept_0_alpha -1.343862  0.643686 -2.0878   0.03608 *  
#> tanktemp_1_alpha          0         0       0         0    
#> gastemp_2_alpha    0.056737  0.011170  5.0792 3.714e-07 ***
#> tankpres_3_alpha          0         0       0         0    
#> gaspres_4_alpha           0         0       0         0    
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> Penalized Likelihood:
#> [1] -308.2184